Adds user ability to sign in and sign up
Modal appears on page load, user can click between sign in a sign up,
showing and hiding the appropriate form.
AJAX calls to the server successfully add a user to the db and sign in
supplies the user with a token.
On successful sign up a user is asked to sign in, and automatically
switches to the sign in form.
A successful sign in dismisses the modal.
Next step is to implement change pass and sign out.