Adds User ability to edit plots (w/ bug)
User can edit individual attributes of a plot from the onePlotModal.
Each attribute has it's own form, and on clicking confirm,
a PATCH AJAX call is made to the db, successfully updating the plot.
The ui response is to re-fetch the plot, but all plotEventHandlers
seem to be disabled, not allowing for continued editing without exiting
the current onePlotModal. Will have to reconsider the ui response.