Adds event & api chain to remove plants from plots
A user can remove a single plant instance from a plot by clicking on the
X next to a plant name in the individual plot view. Utilizes the same ui
pattern as other update / delete user actions: Re-fetching the plots and
plant data from the api and closing the modal.